requirements.txt
evasdk
pyzmq
pymongo
Django

-e ./DPT/Control/dpt_module